🚚 Move `mapToConstant` arbitrary into /arbitrary
